Add .zshrc, installation for it and oh-my-zsh + source code pro font

+++ setup.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+# Assumes emacs, vim, and zsh are already installed.
+
 set -e
 
 function get_abs_filename() {
@@ -7,8 +9,14 @@ function get_abs_filename() {
     echo "$(cd "$(dirname "$rel_filename")" && pwd)/$(basename "$rel_filename")"
 }
 
+# Install oh-my-zsh, if not already installed
+if [[ ! -d ~/.oh-my-zsh ]]; then
+    sh -c "$(wget https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh -O -)"
+    rm ~/.zshrc
+fi
+
 # Setup dotfiles
-for dotfile in .gitignore_global .gvimrc .vimrc .spacemacs .vim .ipyrc; do
+for dotfile in .gitignore_global .gvimrc .vimrc .spacemacs .vim .ipyrc .zshrc; do
     if [[ ! -e "../$dotfile" ]]; then
         ln -s $(get_abs_filename "$dotfile") "$HOME/$dotfile"
 
@@ -29,7 +37,21 @@ pushd snippets > /dev/null
 for snippet_dir in $(ls .); do
     abs_snippet_dir=$(get_abs_filename "$snippet_dir")
     pushd "$base_snippet_dir" > /dev/null
-    ln -s "$abs_snippet_dir" "$snippet_dir"
+    if [[ ! -L "$snippet_dir" ]]; then
+        ln -s "$abs_snippet_dir" "$snippet_dir"
+    fi
     popd > /dev/null
 done;
 popd > /dev/null
+
+# Install source code pro font for spacemacs
+if [[ ! -f ~/.fonts/SourceCodePro-BlackIt.otf ]]; then
+    mkdir -p ~/.fonts
+    wget https://github.com/adobe-fonts/source-code-pro/archive/2.010R-ro/1.030R-it.zip -O ~/.fonts/source-code-pro.zip
+    pushd ~/.fonts >/dev/null
+    unzip source-code-pro.zip
+    rm source-code-pro.zip
+    cp source-code-pro-2.010R-ro-1.030R-it/OTF/*.otf .
+    rm -r source-code-pro-2.010R-ro-1.030R-it/
+    popd > /dev/null
+fi
